+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 * Fixed assert failure with 256 byte sub-blocks
+* Made Header / LogicalScreenDesc not optional in Preamble
 
 ## [0.1.0] - 2019-04-25
 * Initial version